tweaked the resize feature: when the display is resized and the user has previously adjusted the size of a TinyMCE editor, which was set up with PERCENTAGE based width (i.e. relative to the rest of the page layout, which will change when you resize the display) the manually configured size will be RESET to the original percentage-based size as soon as the display is SMALLER than the current TinyMCE editor size.
The reason for this edit is that it is hard (needs extra interface bit somewhere) to reset the user resize within the same session (the size is stored in a cookie, so reloading a page keeps the editor size as-is). Normally this is not an issue, but it is when you shrink the browser viewport to be quite a bit smaller than the editor size: with the percentage-based width setup, the editor would otherwise have scaled nicely with this viewport change, but not so after a single user resize (i.e. dragging on the resize triangle in the lower right corner of the TinyMCE editor).

When the editor is wider than the viewport, some of the toolbar button disappear from view and using them requires scrolling the window/viewport. This is a hassle, so now the editor detects when the viewport is resized (shrunk) so that it clips the editor: when this happens the editor width is reset to the original value (which is a percentage value, e.g. width:100%, because this 'reset' only happens for such percentage-based setups).

Try this with the examples:

- resize one or more editors by dragging their resize triangle.

- resize the viewport (browser window size) and notice that when the window partly clips the editor, the percentage size is popping back in: the width has been reset then. (the example 'skins' has <textarea> items with width:80%, so this new behaviour applies there.

- when you change the width: style for one or more of the <textarea>s in the 'skins' example, you'll note that any resize is permanent; the editor is never reset to the original width here.